requirements.txt
Django==4.0.5 
djangorestframework==3.13.1
psycopg2==2.9.3
celery==5.2.7
redis==4.3.4
gunicorn==20.1.0